+++ src/ingest_validation_tools/schema_loader.py
@@ -251,10 +251,7 @@ def get_table_schema(
         if isinstance(field, str):
             continue
         _add_level_1_description(field)
-        # TODO: Re-enable when all assay names used in the schemas are recognized
-        #       by the assay service, and the list in enums.py has been uncommented.
-        #       https://github.com/hubmapconsortium/ingest-validation-tools/issues/1023
-        # _validate_level_1_enum(field)
+        _validate_level_1_enum(field)
 
         _add_constraints(field, optional_fields, offline=offline, names=names)
         _validate_field(field)