+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## v0.0.14 - in progress
+- Return directory schema version and refactor.
 - Add Clinical Imaging schemas.
 - Test under both Python 3.6 and Python 3.10
 - Cosmetic updates to the Slide-seq directory structure schema.