+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 ## v0.0.14 - in progress
 - Updated IMS directory structure schema.
+- Add Clinical Imaging schemas.
+- Test under both Python 3.6 and Python 3.10
+- Cosmetic updates to the Slide-seq directory structure schema.
+- Fix rendering bug on CODEX page by adding linebreaks.
 - Add type hints.
 - Implement versioning for directory schemas.
 - After failure, explain how to continue testing from last error.