+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## v0.0.14 - in progress
 - Return directory schema version and refactor.
+- Updated IMS directory structure schema.
 - Add Clinical Imaging schemas.
 - Test under both Python 3.6 and Python 3.10
 - Cosmetic updates to the Slide-seq directory structure schema.
@@ -37,6 +38,7 @@
 - Temporarily disable checking the assay names in schemas against the global list.
   Entries in the global list are now commented out, and Joel will progressively
   uncomment them.
+- Moved `dataset.json` to `raw` or `src_*` directory for CODEX datasets.
 
 ## v0.0.13 - 2022-01-07
 - Make more fields explicitly numeric