+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## v0.0.14 - in progress
+- Deprecate older LCMS schemas.
 - Updated IMS directory structure schema.
 - Add Clinical Imaging schemas.
 - Test under both Python 3.6 and Python 3.10