+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## v0.0.14 - in progress
+- Add "CODEX2" assay type.
 - Remove HTML reporting options.
 - Updated IMS directory structure schema.
 - Add Clinical Imaging schemas.