+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## v0.0.14 - in progress
+- Mods to plugin validator to fix import problems.
 - Updated LC-MS directory structure schema.
 - Remove HTML reporting options.
 - Updated IMS directory structure schema.