+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## v0.0.14 - in progress
 - Add "CODEX2" assay type.
+- Updated LC-MS directory structure schema.
 - Remove HTML reporting options.
 - Updated IMS directory structure schema.
 - Add Clinical Imaging schemas.